add support tor and i2p
