+++ src/app/settings.js
@@ -106,6 +106,7 @@ Settings.trackers = {
 
 // API Servers
 Settings.apiServer = '';
+Settings.proxyServer = '';
 
 // User interface
 Settings.language = "";
